Enable GitHub's "Approve and run" button to appear on pull-requests from forks
We used a push trigger because that meant we could push the commits that were in the fork into our repo and it would cause the workflow to run on the commits.
That was a good solution when GitHub didn't have a good way to run workflows on forks, but now GitHub has the "Approve and run" (https://docs.github.com/en/actions/managing-workflow-runs/approving-workflow-runs-from-public-forks) feature for pull-requests from forks.

Moving to their system means we only need to press a button on the pull-request to run the workflows instead of adding the forked repo as a remote, pulling down their branch with the commits and then pushing those commits to our remote.